+++ src/main/java/com/debateseason_backend_v1/domain/profile/infrastructure/ProfileJpaRepository.java
@@ -1,0 +1,16 @@
+package com.debateseason_backend_v1.domain.profile.infrastructure;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProfileJpaRepository extends JpaRepository<ProfileEntity, Long> {
+
+	boolean existsByUserId(Long userId);
+
+	boolean existsByNickname(String nickname);
+
+	Optional<ProfileEntity> findByUserId(Long userId);
+}
