+++ src/test/java/com/debateseason_backend_v1/domain/repository/entity/ProfileTest.java
@@ -11,6 +11,7 @@ import com.debateseason_backend_v1.domain.profile.enums.CommunityType;
 import com.debateseason_backend_v1.domain.profile.enums.DistrictType;
 import com.debateseason_backend_v1.domain.profile.enums.GenderType;
 import com.debateseason_backend_v1.domain.profile.enums.ProvinceType;
+import com.debateseason_backend_v1.domain.profile.infrastructure.ProfileEntity;
 
 class ProfileTest {
 
@@ -26,7 +27,7 @@ class ProfileTest {
 		AgeRangeType ageRange = AgeRangeType.TWENTIES;
 
 		// when
-		Profile profile = Profile.builder()
+		ProfileEntity profile = ProfileEntity.builder()
 			.userId(userId)
 			.profileImage(profileColor)
 			.nickname(nickname)
@@ -48,7 +49,7 @@ class ProfileTest {
 	@DisplayName("프로필을 업데이트할 수 있다")
 	void updateProfile() {
 		// given
-		Profile profile = Profile.builder()
+		ProfileEntity profile = ProfileEntity.builder()
 			.userId(1L)
 			.profileImage("RED")
 			.nickname("토론왕")
@@ -82,7 +83,7 @@ class ProfileTest {
 	@DisplayName("프로필을 익명화할 수 있다")
 	void anonymizeProfile() {
 		// given
-		Profile profile = Profile.builder()
+		ProfileEntity profile = ProfileEntity.builder()
 			.userId(1L)
 			.profileImage("RED")
 			.nickname("토론왕")
@@ -108,7 +109,7 @@ class ProfileTest {
 	void getCommunityType() {
 		// given
 		Long communityId = 1L; // DC_INSIDE
-		Profile profile = Profile.builder()
+		ProfileEntity profile = ProfileEntity.builder()
 			.userId(1L)
 			.profileImage("RED")
 			.nickname("토론왕")
@@ -132,7 +133,7 @@ class ProfileTest {
 	@DisplayName("communityId가 null이면 getCommunityType도 null을 반환한다")
 	void getCommunityTypeReturnsNullForNullCommunityId() {
 		// given
-		Profile profile = Profile.builder()
+		ProfileEntity profile = ProfileEntity.builder()
 			.userId(1L)
 			.profileImage("RED")
 			.nickname("토론왕")
